Image cleanup through dithering

+++ openGeeL/renderer/utility/framebuffer.cpp
@@ -90,7 +90,7 @@ namespace geeL {
 				glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB16F, width, height, 0, GL_RGB, GL_FLOAT, 0);
 				break;
 			case RGBA16:
-				glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA16F, width, height, 0, GL_RGBA, GL_FLOAT, 0);
+				glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, width, height, 0, GL_RGBA, GL_FLOAT, 0);
 				break;
 		}
